refactor Frame
